requirements.txt
azure-cognitiveservices-vision-customvision
matplotlib
numpy
streamlit
openpyxl
dotenv
